Ver 1.03 Validation

+++ src/main/java/com/example/springexercise3boot/controllers/AdminController.java
@@ -3,12 +3,17 @@ package com.example.springexercise3boot.controllers;
 import com.example.springexercise3boot.dto.UserProfileDTO;
 import com.example.springexercise3boot.models.UserProfile;
 import com.example.springexercise3boot.services.UserProfileService;
+import com.example.springexercise3boot.util.UserProfileValidator;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindException;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.view.RedirectView;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @RestController
@@ -16,10 +21,12 @@ import java.util.List;
 @Slf4j
 public class AdminController {
     private final UserProfileService userProfileService;
+    private final UserProfileValidator userProfileValidator;
 
     @Autowired
-    public AdminController(UserProfileService userProfileService) {
+    public AdminController(UserProfileService userProfileService, UserProfileValidator userProfileValidator) {
         this.userProfileService = userProfileService;
+        this.userProfileValidator = userProfileValidator;
     }
 
     @GetMapping("/index")
@@ -44,37 +51,35 @@ public class AdminController {
     }
 
     @PostMapping("/addUser")
-    public RedirectView addUser(@ModelAttribute UserProfileDTO profileDTO) {
-        log.info("Inserting user into database. Login: " + profileDTO.getUsername()
+    public ResponseEntity<String> addUser(@Valid UserProfileDTO profileDTO, BindingResult bindingResult)
+            throws BindException {
+        log.info("Attempt inserting user into database. Login: " + profileDTO.getUsername()
                 + " Role: " + profileDTO.getRole()
                 + " Name: " + profileDTO.getName()
                 + " Email: " + profileDTO.getEmail());
 
-        if (profileDTO.getUsername().equals("")) {
-            return new RedirectView("/");
-        } else if (userProfileService.findByUsername(profileDTO.getUsername()) != null) {
-            return new RedirectView("/admin/userExists");
+        userProfileValidator.validate(profileDTO, bindingResult);
+
+        if (bindingResult.hasErrors()) {
+            log.error("User can't be added because of invalid data");
+            throw new BindException(bindingResult);
         }
 
+        log.info("User successfully created");
         userProfileService.save(convertToUserProfile(profileDTO));
 
-        return new RedirectView("/admin/index");
-    }
-
-    @GetMapping("/userExists")
-    public ModelAndView showUserExistsMessage() {
-        log.info("Showing \"User already exist\" page");
-
-        return new ModelAndView("admin/login-exists");
+        return new ResponseEntity<>("User successfully created", HttpStatus.CREATED);
     }
 
     @GetMapping("/deleteUser")
-    public RedirectView deleteUser(@RequestParam("userIdToDelete") int id) {
-        log.info("Removing user from database, ID: " + id);
+    public ResponseEntity<String> deleteUser(@RequestParam("userIdToDelete") int id) {
+        log.info("Attempt removing user from database, ID: " + id);
 
-        userProfileService.delete(id);
-
-        return new RedirectView("/admin/index");
+        if (userProfileService.findOne(id) != null) {
+            userProfileService.delete(id);
+        }
+        log.info("User successfully deleted");
+        return new ResponseEntity<>("User successfully deleted", HttpStatus.OK);
     }
 
     @GetMapping("/updateUserForm")
@@ -92,12 +97,21 @@ public class AdminController {
     }
 
     @PostMapping("/updateUser")
-    public RedirectView updateUser(@ModelAttribute UserProfileDTO profileDTO) {
-        log.info("Updating user in database. Name: " + profileDTO.getName() + " Email: " + profileDTO.getEmail());
+    public ResponseEntity<String> updateUser(@Valid UserProfileDTO profileDTO, BindingResult bindingResult)
+            throws BindException {
+        log.info("Attempt to update user in database. Name: " + profileDTO.getName() + " Email: " + profileDTO.getEmail());
 
+        userProfileValidator.validate(profileDTO, bindingResult);
+
+        if (bindingResult.hasErrors()) {
+            log.error("User can't be updated because of invalid data");
+            throw new BindException(bindingResult);
+        }
+
+        log.info("User successfully updated");
         userProfileService.update(profileDTO.getId(), convertToUserProfile(profileDTO));
 
-        return new RedirectView("/admin/index");
+        return new ResponseEntity<>("User successfully updated", HttpStatus.OK);
     }
 
     private UserProfile convertToUserProfile(UserProfileDTO profileDTO) {
